improved polybar launch to accomodate for situation when only external monitor is active

+++ polybar/launch.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env sh
 
+if ! type "xrandr" &> /dev/null ; then
+  echo "showing polybar only for if X-11 is present"
+  exit 1
+fi
+
 # Terminate already running bar instances
 killall -q polybar
 
@@ -9,18 +14,25 @@ while pgrep -u $UID -x polybar >/dev/null; do sleep 1; done
 lan_interface=$(ip link show | awk -F ':' '{print $2}' | grep 'en' | xargs)
 wlan_interace=$(ip link show | awk -F ':' '{print $2}' | grep 'wl' | xargs)
 
-# Launch polybar named example
-if type "xrandr"; then
-  for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
-    if [[ "$m" == *"HDMI"* ]]; then
-      MONITOR=$m LAN_INTERFACE=$lan_interface WLAN_INTERFACE=$wlan_interace polybar --reload secondary &
-    fi
-    if [[ "$m" == *"eDP"* ]]; then
-      MONITOR=$m LAN_INTERFACE=$lan_interface WLAN_INTERFACE=$wlan_interace polybar --reload main &
-    fi
-  done
-else
-  polybar --reload example &
+monitors=($(xrandr --query | grep " connected" | cut -d" " -f1))
+monitorsNumber=${#monitors[@]}
+
+# if there is only one output, show main status bar
+if [ $monitorsNumber == 1 ]; then
+  MONITOR=$m LAN_INTERFACE=$lan_interface WLAN_INTERFACE=$wlan_interace polybar --reload main &
+  exit 0
 fi
 
-echo "Bars launched..."
+# if there is more than one output, show main status bar on laptop screen
+# and secondary status bars on others
+for (( i=0; i<${monitorsNumber}; i++ ));
+do
+  m=${monitors[$i]}
+  if [[ "$m" == *"eDP"* ]]; then
+    MONITOR=$m LAN_INTERFACE=$lan_interface WLAN_INTERFACE=$wlan_interace polybar --reload main &
+  else
+    MONITOR=$m LAN_INTERFACE=$lan_interface WLAN_INTERFACE=$wlan_interace polybar --reload secondary &
+  fi
+done
+
+exit 0
